+++ Modules/SharePointDsc/DSCResources/MSFT_SPAccessServiceApp/MSFT_SPAccessServiceApp.psm1
@@ -45,7 +45,8 @@ function Get-TargetResource
             return $nullReturn 
         }
         $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "Access Services Web Service Application" 
+            $_.TypeName -eq "Access Services Web Service Application" -or `
+            $_.GetType().FullName -eq "Microsoft.Office.Access.Services.MossHost.AccessServicesWebServiceApplication"            
         }
 
         if ($null -eq $serviceApp) 
@@ -120,7 +121,8 @@ function Set-TargetResource
             $params = $args[0]
             
             $app = Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript { 
-                $_.TypeName -eq "Access Services Web Service Application"  
+                $_.TypeName -eq "Access Services Web Service Application" -or `
+                $_.GetType().FullName -eq "Microsoft.Office.Access.Services.MossHost.AccessServicesWebServiceApplication"
             }
             Remove-SPServiceApplication $app -Confirm:$false
         }