+++ Modules/SharePointDsc/DSCResources/MSFT_SPAppManagementServiceApp/MSFT_SPAppManagementServiceApp.psm1
@@ -53,7 +53,8 @@ function Get-TargetResource
             return $nullReturn
         }
         $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "App Management Service Application" 
+            $_.TypeName -eq "App Management Service Application" -or `
+            $_.GetType().FullName -eq "Microsoft.SharePoint.AppManagement.AppManagementServiceApplication"            
         }
 
         if ($null -eq $serviceApp) 
@@ -177,7 +178,8 @@ function Set-TargetResource
                 $appPool = Get-SPServiceApplicationPool -Identity $params.ApplicationPool
                 
                 $app = Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript { 
-                    $_.TypeName -eq "App Management Service Application"  
+                    $_.TypeName -eq "App Management Service Application" -or `
+                    $_.GetType().FullName -eq "Microsoft.SharePoint.AppManagement.AppManagementServiceApplication"   
                 }
                 $app.ApplicationPool = $appPool
                 $app.Update()
@@ -195,7 +197,8 @@ function Set-TargetResource
             $params = $args[0]
             
             $app = Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript { 
-                $_.TypeName -eq "App Management Service Application"
+                $_.TypeName -eq "App Management Service Application" -or `
+                $_.GetType().FullName -eq "Microsoft.SharePoint.AppManagement.AppManagementServiceApplication"   
             }
             Remove-SPServiceApplication $app -Confirm:$false
         }