+++ Modules/SharePointDsc/DSCResources/MSFT_SPExcelServiceApp/MSFT_SPExcelServiceApp.psm1
@@ -50,7 +50,8 @@ function Get-TargetResource
             return $nullReturn 
         }
         $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "Excel Services Application Web Service Application" 
+            $_.TypeName -eq "Excel Services Application Web Service Application" -or `
+            $_.GetType().FullName -eq "Microsoft.Office.Excel.Server.MossHost.ExcelServerWebServiceApplication"    
         }
 
         if ($null -eq $serviceApp) 
@@ -125,7 +126,8 @@ function Set-TargetResource
             $params = $args[0]
             
             $appService =  Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript {
-                    $_.TypeName -eq "Excel Services Application Web Service Application"  
+                $_.TypeName -eq "Excel Services Application Web Service Application" -or `
+                $_.GetType().FullName -eq "Microsoft.Office.Excel.Server.MossHost.ExcelServerWebServiceApplication"  
             }
             Remove-SPServiceApplication $appService -Confirm:$false
         }