+++ Modules/SharePointDsc/DSCResources/MSFT_SPPerformancePointServiceApp/MSFT_SPPerformancePointServiceApp.psm1
@@ -52,8 +52,9 @@ function Get-TargetResource
         { 
             return $nullReturn 
         }
-        $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "PerformancePoint Service Application" 
+        $serviceApp = $serviceApps | Where-Object -FilterScript {
+            $_.TypeName -eq "PerformancePoint Service Application" -or `
+            $_.GetType().FullName -eq "Microsoft.PerformancePoint.Scorecards.BIMonitoringServiceApplication"
         }
 
         if ($null -eq $serviceApp) 
@@ -172,7 +173,8 @@ function Set-TargetResource
 
                 Get-SPServiceApplication -Name $params.Name `
                     | Where-Object -FilterScript { 
-                        $_.TypeName -eq "PerformancePoint Service Application" 
+                        $_.TypeName -eq "PerformancePoint Service Application" -or `
+                        $_.GetType().FullName -eq "Microsoft.PerformancePoint.Scorecards.BIMonitoringServiceApplication"
                     } | Set-SPPerformancePointServiceApplication -ApplicationPool $appPool
             }
         }
@@ -186,7 +188,8 @@ function Set-TargetResource
             $params = $args[0]
             
             $app = Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript {
-                $_.TypeName -eq "PerformancePoint Service Application"  
+                $_.TypeName -eq "PerformancePoint Service Application" -or `
+                $_.GetType().FullName -eq "Microsoft.PerformancePoint.Scorecards.BIMonitoringServiceApplication"
             }
             Remove-SPServiceApplication $app -Confirm:$false
         }