+++ Modules/SharePointDsc/DSCResources/MSFT_SPSearchServiceApp/MSFT_SPSearchServiceApp.psm1
@@ -78,7 +78,8 @@ function Get-TargetResource
             return $nullReturn 
         }
         $serviceApp = $serviceApps | Where-Object -FilterScript { 
-            $_.TypeName -eq "Search Service Application" 
+            $_.TypeName -eq "Search Service Application" -or `
+            $_.GetType().FullName -eq "Microsoft.Office.Server.Search.Administration.SearchServiceApplication"
         }
 
         if ($null -eq $serviceApp) 
@@ -254,7 +255,8 @@ function Set-TargetResource
                 {
                     $serviceApp = Get-SPServiceApplication -Name $params.Name | `
                         Where-Object -FilterScript { 
-                            $_.TypeName -eq "Search Service Application" 
+                            $_.TypeName -eq "Search Service Application" -or `
+                            $_.GetType().FullName -eq "Microsoft.Office.Server.Search.Administration.SearchServiceApplication"
                         }
                     $serviceApp.SearchCenterUrl = $params.SearchCenterUrl
                     $serviceApp.Update()
@@ -274,7 +276,8 @@ function Set-TargetResource
             
             $serviceApp = Get-SPServiceApplication -Name $params.Name | `
                 Where-Object -FilterScript { 
-                    $_.TypeName -eq "Search Service Application" 
+                    $_.TypeName -eq "Search Service Application" -or `
+                    $_.GetType().FullName -eq "Microsoft.Office.Server.Search.Administration.SearchServiceApplication"
                 }
             $appPool = Get-SPServiceApplicationPool -Identity $params.ApplicationPool
             $setParams = @{
@@ -293,7 +296,8 @@ function Set-TargetResource
             {
                 $serviceApp = Get-SPServiceApplication -Name $params.Name | `
                     Where-Object -FilterScript { 
-                        $_.TypeName -eq "Search Service Application" 
+                        $_.TypeName -eq "Search Service Application" -or `
+                        $_.GetType().FullName -eq "Microsoft.Office.Server.Search.Administration.SearchServiceApplication" 
                     }
                 $serviceApp.SearchCenterUrl = $params.SearchCenterUrl
                 $serviceApp.Update()
@@ -311,7 +315,8 @@ function Set-TargetResource
             $params = $args[0]
             
             $serviceApp =  Get-SPServiceApplication -Name $params.Name | Where-Object -FilterScript {
-                $_.TypeName -eq "Search Service Application"
+                $_.TypeName -eq "Search Service Application" -or `
+                $_.GetType().FullName -eq "Microsoft.Office.Server.Search.Administration.SearchServiceApplication"
             }
             Remove-SPServiceApplication $serviceApp -Confirm:$false
         }